renamed files and imports according to naming convention
