+++ helpers/auth0Helper.js
