+++ routes/api/users.js
@@ -2,7 +2,7 @@ const express = require('express');
 const router = express.Router();
 const { getAllUsers } = require('../../controllers/userController');
 const logToken  = require('../../middleware/logTokenMiddleware');
-const checkPermissions = require('../../middleware/permissionsMiddleware');
+const checkPermissions = require('../../middleware/permissionMiddleware');
 /**
  * @swagger
  * /users:
